+++ data/colors/summer.lua
@@ -1,28 +1,34 @@
+local config = require "core.config"
 local style = require "core.style"
-local common = require "core.common"
 
-style.background = { common.color "#fbfbfb" }
-style.background2 = { common.color "#f2f2f2" }
-style.background3 = { common.color "#f2f2f2" }
-style.text = { common.color "#404040" }
-style.caret = { common.color "#fc1785" }
-style.accent = { common.color "#fc1785" }
-style.dim = { common.color "#b0b0b0" }
-style.divider = { common.color "#e8e8e8" }
-style.selection = { common.color "#b7dce8" }
-style.line_number = { common.color "#d0d0d0" }
-style.line_number2 = { common.color "#808080" }
-style.line_highlight = { common.color "#f2f2f2" }
-style.scrollbar = { common.color "#e0e0e0" }
-style.scrollbar2 = { common.color "#c0c0c0" }
+style.register_theme("summer", {
+  background = "#fbfbfb",
+  background2 = "#f2f2f2",
+  background3 = "#f2f2f2",
+  text = "#404040",
+  caret = "#fc1785",
+  accent = "#fc1785",
+  dim = "#b0b0b0",
+  divider = "#e8e8e8",
+  selection = "#b7dce8",
+  line_number = "#d0d0d0",
+  line_number2 = "#808080",
+  line_highlight = "#f2f2f2",
+  scrollbar = "#e0e0e0",
+  scrollbar2 = "#c0c0c0",
+  syntax = {
+    normal = "#181818",
+    symbol = "#181818",
+    comment = "#22a21f",
+    keyword = "#fb6620",
+    keyword2 = "#fc1785",
+    number = "#1586d2",
+    literal = "#1586d2",
+    string = "#1586d2",
+    operator = "#fb6620",
+    ["function"] = "#fc1785",
+  },
+})
 
-style.syntax["normal"] = { common.color "#181818" }
-style.syntax["symbol"] = { common.color "#181818" }
-style.syntax["comment"] = { common.color "#22a21f" }
-style.syntax["keyword"] = { common.color "#fb6620" }
-style.syntax["keyword2"] = { common.color "#fc1785" }
-style.syntax["number"] = { common.color "#1586d2" }
-style.syntax["literal"] = { common.color "#1586d2" }
-style.syntax["string"] = { common.color "#1586d2" }
-style.syntax["operator"] = { common.color "#fb6620" }
-style.syntax["function"] = { common.color "#fc1785" }
+config.theme = "summer"
+return style