+++ data/core/logview.lua
@@ -2,6 +2,7 @@ local core    = require "core"
 local common  = require "core.common"
 local config  = require "core.config"
 local command = require "core.command"
+local ContextMenu = require "core.contextmenu"
 local keymap  = require "core.keymap"
 local style   = require "core.style"
 local View    = require "core.view"
@@ -20,16 +21,20 @@ end
 local item_height_result = {}
 
 
-local function get_item_height(item)
-  local h = item_height_result[item]
+local function get_item_height(item, font)
+  local cache = item_height_result[item]
+  local key = font:get_size()
+  local h = cache and cache[key]
   if not h then
     h = {}
     local l = 1 + lines(item.text) + lines(item.info or "")
-    h.normal = style.font:get_height() + style.padding.y
-    h.expanded = l * style.font:get_height() + style.padding.y
+    h.normal = font:get_height() + style.padding.y
+    h.expanded = l * font:get_height() + style.padding.y
     h.current = h.normal
     h.target = h.current
-    item_height_result[item] = h
+    cache = cache or {}
+    cache[key] = h
+    item_height_result[item] = cache
   end
   return h
 end
@@ -48,6 +53,7 @@ function LogView:new()
   self.scrollable = true
   self.yoffset = 0
   self.selected = nil
+  self.font_delta = 0
 
   core.status_view:show_message("i", style.text,
     "click to select  \xc2\xb7  ctrl+c to copy entry  \xc2\xb7  ctrl+a to copy all")
@@ -59,26 +65,55 @@ function LogView:get_name()
 end
 
 
-local function is_expanded(item)
-  local item_height = get_item_height(item)
+function LogView:get_log_font()
+  return self.font_delta == 0 and style.font or style.font:copy(style.font:get_size() + self.font_delta)
+end
+
+
+function LogView:get_log_icon_font()
+  return self.font_delta == 0 and style.icon_font or style.icon_font:copy(style.icon_font:get_size() + self.font_delta)
+end
+
+
+function LogView:invalidate_item_heights()
+  item_height_result = {}
+  self.expanding = {}
+end
+
+
+function LogView:change_font_size(delta)
+  self.font_delta = math.max(-8, math.min(24, self.font_delta + delta))
+  self:invalidate_item_heights()
+end
+
+
+function LogView:reset_font_size()
+  self.font_delta = 0
+  self:invalidate_item_heights()
+end
+
+
+local function is_expanded(item, font)
+  local item_height = get_item_height(item, font)
   return item_height.target == item_height.expanded
 end
 
 
 function LogView:expand_item(item)
-  item = get_item_height(item)
+  item = get_item_height(item, self:get_log_font())
   item.target = item.target == item.expanded and item.normal or item.expanded
   table.insert(self.expanding, item)
 end
 
 
 function LogView:each_item()
+  local font = self:get_log_font()
   local x, y = self:get_content_offset()
   y = y + style.padding.y + self.yoffset
   return coroutine.wrap(function()
     for i = #core.log_items, 1, -1 do
       local item = core.log_items[i]
-      local h = get_item_height(item).current
+      local h = get_item_height(item, font).current
       coroutine.yield(i, item, x, y, self.size.x, h)
       y = y + h
     end
@@ -115,12 +150,16 @@ function LogView:on_mouse_pressed(button, px, py, clicks)
 
   if selected then
     self.selected = selected
-    if keymap.modkeys["ctrl"] then
+    if button == "right" then
+      command.perform("context-menu:show", px, py)
+    elseif keymap.modkeys["ctrl"] then
       system.set_clipboard(core.get_log(selected))
       core.status_view:show_message("i", style.text, "copied entry #"..index.." to clipboard")
     else
       self:expand_item(selected)
     end
+  elseif button == "right" then
+    command.perform("context-menu:show", px, py)
   end
 
   return true
@@ -163,7 +202,7 @@ local function draw_text_multiline(font, text, x, y, color)
   local th = font:get_height()
   local resx = x
   for line in text:gmatch("[^\n]+") do
-    resx = renderer.draw_text(style.font, line, x, y, color)
+    resx = renderer.draw_text(font, line, x, y, color)
     y = y + th
   end
   return resx, y
@@ -174,14 +213,16 @@ local datestr = os.date()
 function LogView:draw()
   self:draw_background(style.background)
 
-  local th = style.font:get_height()
+  local font = self:get_log_font()
+  local icon_font = self:get_log_icon_font()
+  local th = font:get_height()
   local lh = th + style.padding.y -- for one line
   local iw = math.max(
-    style.icon_font:get_width(style.log.ERROR.icon),
-    style.icon_font:get_width(style.log.INFO.icon)
+    icon_font:get_width(style.log.ERROR.icon),
+    icon_font:get_width(style.log.INFO.icon)
   )
 
-  local tw = style.font:get_width(datestr)
+  local tw = font:get_width(datestr)
   for _, item, x, y, w, h in self:each_item() do
     if y + h >= self.position.y and y <= self.position.y + self.size.y then
       if item == self.selected then
@@ -191,7 +232,7 @@ function LogView:draw()
       x = x + style.padding.x
 
       x = common.draw_text(
-        style.icon_font,
+        icon_font,
         style.log[item.level].color,
         style.log[item.level].icon,
         "center",
@@ -201,33 +242,44 @@ function LogView:draw()
 
       -- timestamps are always 15% of the width
       local time = os.date(nil, item.time)
-      common.draw_text(style.font, style.dim, time, "left", x, y, tw, lh)
+      common.draw_text(font, style.dim, time, "left", x, y, tw, lh)
       x = x + tw + style.padding.x
 
       w = w - (x - self:get_content_offset())
 
-      if is_expanded(item) then
+      if is_expanded(item, font) then
         y = y + common.round(style.padding.y / 2)
-        _, y = draw_text_multiline(style.font, item.text, x, y, style.text)
+        _, y = draw_text_multiline(font, item.text, x, y, style.text)
 
         local at = "at " .. common.home_encode(item.at)
-        _, y = common.draw_text(style.font, style.dim, at, "left", x, y, w, lh)
+        _, y = common.draw_text(font, style.dim, at, "left", x, y, w, lh)
 
         if item.info then
-          _, y = draw_text_multiline(style.font, item.info, x, y, style.dim)
+          _, y = draw_text_multiline(font, item.info, x, y, style.dim)
         end
       else
         local line, has_newline = string.match(item.text, "([^\n]+)(\n?)")
         if has_newline ~= "" then
           line = line .. " ..."
         end
-        _, y = common.draw_text(style.font, style.text, line, "left", x, y, w, lh)
+        _, y = common.draw_text(font, style.text, line, "left", x, y, w, lh)
       end
 
       core.pop_clip_rect()
     end
   end
   LogView.super.draw_scrollbar(self)
+end
+
+
+function LogView:on_context_menu()
+  return { items = {
+    { text = "Copy", command = "log:copy-entry" },
+    ContextMenu.DIVIDER,
+    { text = "Font Size +", command = "log:font-size-increase" },
+    { text = "Font Size -", command = "log:font-size-decrease" },
+    { text = "Font Reset", command = "log:font-size-reset" },
+  } }, self
 end
 
 local function is_log_view()
@@ -246,6 +298,15 @@ command.add(is_log_view, {
   ["log:copy-all"] = function()
     system.set_clipboard(core.get_log())
     core.status_view:show_message("i", style.text, "copied all log entries to clipboard")
+  end,
+  ["log:font-size-increase"] = function(lv)
+    lv:change_font_size(1 * SCALE)
+  end,
+  ["log:font-size-decrease"] = function(lv)
+    lv:change_font_size(-1 * SCALE)
+  end,
+  ["log:font-size-reset"] = function(lv)
+    lv:reset_font_size()
   end,
 })